moved widget py files into core folder
