+++ app/core/ProjectEstimateWidget.py
