+++ app/data/database/user_tasks.json
@@ -1,1 +1,1 @@
-[{"task_id": 1, "tasks": {"Kitchen": {"Demolition": ["Sink", "Disposal", "Microwave"], "Rebuild": ["Countertops"], "Cosmetic": ["Walls"]}, "Bathroom": {"Demolition": ["Shower", "Walls"], "Rebuild": ["Floor", "Shower"], "Cosmetic": ["Baseboard"]}}}]
+[{"task_id": 1, "tasks": {"Kitchen": {"Demolition": ["Sink", "Disposal", "Microwave"], "Rebuild": ["Countertops"], "Cosmetic": ["Walls"]}, "Bathroom": {"Demolition": ["Shower", "Walls"], "Rebuild": ["Floor", "Shower"], "Cosmetic": ["Baseboard"]}}}, {"task_id": 2, "tasks": {"Kitchen": {"Demolition": ["Faucet", "Sink", "Refrigerator"], "Rebuild": [], "Cosmetic": []}, "Bathroom": {"Demolition": ["Sink", "Shower", "Walls"], "Rebuild": [], "Cosmetic": []}}}]
