--- tasks_kw.json
@@ -1,1 +1,0 @@
-{"Kitchen": {"Demolition": ["Cabinets", "Countertops", "Walls", "Flooring", "Faucet", "Sink", "Disposal", "Refrigerator", "Microwave", "Range", "Dishwasher", "Lights", "Outlets"], "Rebuild": ["Cabinets", "Countertops"], "Cosmetic": ["Ceilings", "Walls", "Baseboards"]}, "Bathroom": {"Demolition": ["Toilet", "Sink", "Shower", "Walls", "Flooring"], "Rebuild": ["Walls", "Floor", "Shower"], "Cosmetic": ["Walls", "Baseboard", "Tiling"]}, "Bedroom": {"Demolition": ["Walls"], "Rebuild": [], "Cosmetic": []}, "Multiple Room": {"Demolition": [], "Rebuild": [], "Cosmetic": []}, "Addition": {"Demolition": [], "Rebuild": [], "Cosmetic": []}, "Other": {"Demolition": [], "Rebuild": [], "Cosmetic": []}}
